pool: impl Drop for Relay
Move `Relay` to `InternalRelay`
Add `Relay` with Drop support
Add `Relay::notifications` method
Move `RelayStatus` to `relay::status` module
