+++ crates/nostr-relay-pool/src/lib.rs
@@ -24,4 +24,6 @@ pub use self::relay::options::{
     RequestAutoCloseOptions, RequestOptions,
 };
 pub use self::relay::stats::RelayConnectionStats;
-pub use self::relay::{ActiveSubscription, InternalSubscriptionId, Relay, RelayStatus};
+pub use self::relay::{
+    ActiveSubscription, InternalSubscriptionId, Relay, RelayNotification, RelayStatus,
+};
