+++ crates/nostr-relay-pool/src/relay/status.rs
@@ -1,0 +1,47 @@
+// Copyright (c) 2022-2023 Yuki Kishimoto
+// Copyright (c) 2023-2024 Rust Nostr Developers
+// Distributed under the MIT software license
+
+//! Relay status
+
+use core::fmt;
+
+/// Relay connection status
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub enum RelayStatus {
+    /// Relay initialized
+    Initialized,
+    /// Pending
+    Pending,
+    /// Connecting
+    Connecting,
+    /// Relay connected
+    Connected,
+    /// Relay disconnected, will retry to connect again
+    Disconnected,
+    /// Stop
+    Stopped,
+    /// Relay completely disconnected
+    Terminated,
+}
+
+impl fmt::Display for RelayStatus {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Initialized => write!(f, "Initialized"),
+            Self::Pending => write!(f, "Pending"),
+            Self::Connecting => write!(f, "Connecting"),
+            Self::Connected => write!(f, "Connected"),
+            Self::Disconnected => write!(f, "Disconnected"),
+            Self::Stopped => write!(f, "Stopped"),
+            Self::Terminated => write!(f, "Terminated"),
+        }
+    }
+}
+
+impl RelayStatus {
+    /// Check if is `disconnected`, `stopped` or `terminated`
+    pub(crate) fn is_disconnected(&self) -> bool {
+        matches!(self, Self::Disconnected | Self::Stopped | Self::Terminated)
+    }
+}
